Renamed files in submodules and removed global h54s variable
