+++ src/h54s.js
@@ -111,10 +111,10 @@ var h54s = module.exports = function(config) {
 h54s.version = '__version__';
 
 
-h54s.prototype = require('./methods/methods.js');
+h54s.prototype = require('./methods');
 
-h54s.Tables = require('./tables/tables.js');
-h54s.Files = require('./files/files.js');
+h54s.Tables = require('./tables');
+h54s.Files = require('./files');
 h54s.SasData = require('./sasData.js');
 
 //self invoked function module
